+++ eiam-20211201/alibabacloud_eiam20211201/models/_generate_oauth_token_request.py
@@ -11,8 +11,12 @@ class GenerateOauthTokenRequest(DaraModel):
         self,
         application_id: str = None,
         audience: str = None,
+        grant_type: str = None,
         instance_id: str = None,
+        requested_token_type: str = None,
         scope_values: List[str] = None,
+        subject_token: str = None,
+        subject_token_type: str = None,
     ):
         # The application ID.
         # 
@@ -22,14 +26,18 @@ class GenerateOauthTokenRequest(DaraModel):
         # 
         # This parameter is required.
         self.audience = audience
+        self.grant_type = grant_type
         # The instance ID.
         # 
         # This parameter is required.
         self.instance_id = instance_id
+        self.requested_token_type = requested_token_type
         # The permission scopes.
         # 
         # This parameter is required.
         self.scope_values = scope_values
+        self.subject_token = subject_token
+        self.subject_token_type = subject_token_type
 
     def validate(self):
         pass
@@ -45,11 +53,23 @@ class GenerateOauthTokenRequest(DaraModel):
         if self.audience is not None:
             result['Audience'] = self.audience
 
+        if self.grant_type is not None:
+            result['GrantType'] = self.grant_type
+
         if self.instance_id is not None:
             result['InstanceId'] = self.instance_id
 
+        if self.requested_token_type is not None:
+            result['RequestedTokenType'] = self.requested_token_type
+
         if self.scope_values is not None:
             result['ScopeValues'] = self.scope_values
+
+        if self.subject_token is not None:
+            result['SubjectToken'] = self.subject_token
+
+        if self.subject_token_type is not None:
+            result['SubjectTokenType'] = self.subject_token_type
 
         return result
 
@@ -61,11 +81,23 @@ class GenerateOauthTokenRequest(DaraModel):
         if m.get('Audience') is not None:
             self.audience = m.get('Audience')
 
+        if m.get('GrantType') is not None:
+            self.grant_type = m.get('GrantType')
+
         if m.get('InstanceId') is not None:
             self.instance_id = m.get('InstanceId')
 
+        if m.get('RequestedTokenType') is not None:
+            self.requested_token_type = m.get('RequestedTokenType')
+
         if m.get('ScopeValues') is not None:
             self.scope_values = m.get('ScopeValues')
+
+        if m.get('SubjectToken') is not None:
+            self.subject_token = m.get('SubjectToken')
+
+        if m.get('SubjectTokenType') is not None:
+            self.subject_token_type = m.get('SubjectTokenType')
 
         return self